requirements.txt
llama-index-core==0.14.13
llama-index-embeddings-huggingface==0.6.1
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.1
llama-index-llms-openai==0.6.15
llama-index-readers-file==0.5.6
llama-index-workflows==2.13.1
pymupdf
sentence-transformers
transformers
streamlit
huggingface_hub
numpy @ file:///Users/ec2-user/croot/numpy_and_numpy_base_1768376198994/work/dist/numpy-2.4.1-cp311-cp311-macosx_12_0_arm6
ollama==0.6.1
pandas==2.3.3
pikepdf==10.2.0
PyMuPDF==1.26.7
pypdf==6.6.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
